Improve sample route, add some tests
There is now a separation between routes and models and examples
on how to do tests.

DI was reconfigured to autowire all of those classes, but in the
tests we are doing it explicitly. The container could have been
used to inject those dependencies there, but I decided to let
them explicit on the test suites.